Create Paint Area Calculator
You are painting a wall. The instructions on the paint can say that 1 can of paint can cover 5 square meters of wall. Given a random height and width of the wall, calculate how many cans of paint you'll need to buy.

number of cans = (wall height x wall width) ÷ coverage per can.
But because you can't buy 0.6 of a paint can, the result should be rounded to 2 cans.

IMPORTANT: Please note that the name of the function and parameters must match those on line 13 for the code to work.